[Laser] Project points from the events
Summary: Some work done on projecting points

Test Plan: Needs to be tested with the actual 3D points

Reviewers: nand, piotr, hanq, michsien

Reviewed By: michsien

Maniphest Tasks: T75

Differential Revision: https://ec2-52-30-7-90.eu-west-1.compute.amazonaws.com/D153

+++ procam/LaserDetector.cc
@@ -214,7 +214,7 @@ void LaserDetector::detect() {
         dv::Point laser;
         laser.x = track->x;
         laser.y = track->y;
-        laser.depth = baseline_->getDepthImage().at<double>(track->y, track->x);
+        laser.depth = baseline_->getDepthImage().at<float>(track->y, track->x);
         master_->detectedLaser(laser, thriftColor);
       }
     }
